+++ views/forms/signin.ejs
@@ -42,27 +42,30 @@
     </div>
 </div>
 
-<!-- start coding here -->
+<!-- Start coding here -->
 <div class="container d-flex align-items-center justify-content-center min-vh-100">
     <main class="form-signin w-100 m-auto text-center">
         <form>
-            <h1 class="h3 mb-3 fw-normal">Please sign up</h1>
+            <h1 class="h3 mb-3 fw-normal">Please sign in</h1>
 
             <div class="form-floating">
-                <input type="text" class="form-control" id="floatingName" placeholder="Full Name">
-                <label for="floatingName">Full Name</label>
-            </div>
-            <div class="form-floating">
-                <input type="email" class="form-control" id="floatingInput" placeholder="name@example.com">
+                <input type="email" class="form-control" id="floatingInput" placeholder="name@example.com" required autofocus>
                 <label for="floatingInput">Email address</label>
             </div>
             <div class="form-floating">
-                <input type="password" class="form-control" id="floatingPassword" placeholder="Password">
+                <input type="password" class="form-control" id="floatingPassword" placeholder="Password" required>
                 <label for="floatingPassword">Password</label>
             </div>
 
-            <button class="w-100 btn btn-lg btn-primary" type="submit">Sign up</button>
-            <p class="mt-3">Already have an account? <a href="/signin">Sign in</a></p>
+            <div class="form-check text-start my-3">
+                <input class="form-check-input" type="checkbox" value="remember-me" id="rememberMe">
+                <label class="form-check-label" for="rememberMe">
+                    Remember me
+                </label>
+            </div>
+
+            <button class="w-100 btn btn-lg btn-primary" type="submit">Sign in</button>
+            <p class="mt-3">Don't have an account? <a href="/signup">Sign up</a></p>
         </form>
     </main>
 </div>
